feat: add repositories, mappers, exception, and controllers

+++ backend/src/main/java/com/jjapartments/backend/models/MonthlyReport.java
@@ -5,9 +5,9 @@ public class MonthlyReport {
     private int id;
     private int year;
     private int month;
-    private double totalEarnings;
-    private double totalExpenses;
-    private double netIncome;
+    private float totalEarnings;
+    private float totalExpenses;
+    private float netIncome;
 
 
     public int getId() {
@@ -34,27 +34,27 @@ public class MonthlyReport {
         this.month = month;
     }
 
-    public double getTotalEarnings() {
+    public float getTotalEarnings() {
         return this.totalEarnings;
     }
 
-    public void setTotalEarnings(double totalEarnings) {
+    public void setTotalEarnings(float totalEarnings) {
         this.totalEarnings = totalEarnings;
     }
 
-    public double getTotalExpenses() {
+    public float getTotalExpenses() {
         return this.totalExpenses;
     }
 
-    public void setTotalExpenses(double totalExpenses) {
+    public void setTotalExpenses(float totalExpenses) {
         this.totalExpenses = totalExpenses;
     }
 
-    public double getNetIncome() {
+    public float getNetIncome() {
         return this.netIncome;
     }
 
-    public void setNetIncome(double netIncome) {
+    public void setNetIncome(float netIncome) {
         this.netIncome = netIncome;
     }
